Refactor file names and restructure folders
