--- web/src/components/ContactCard2/ContactCard2.stories.js
@@ -1,7 +1,0 @@
-import ContactCard2 from './ContactCard2'
-
-export const generated = () => {
-  return <ContactCard2 />
-}
-
-export default { title: 'Components/ContactCard2' }
