--- web/src/components/ContactCard2/ContactCard2.test.js
@@ -1,11 +1,0 @@
-import { render } from '@redwoodjs/testing'
-
-import ContactCard2 from './ContactCard2'
-
-describe('ContactCard2', () => {
-  it('renders successfully', () => {
-    expect(() => {
-      render(<ContactCard2 />)
-    }).not.toThrow()
-  })
-})
