--- web/src/components/MenuCard2/MenuCard2.stories.js
@@ -1,7 +1,0 @@
-import MenuCard2 from './MenuCard2'
-
-export const generated = () => {
-  return <MenuCard2 />
-}
-
-export default { title: 'Components/MenuCard2' }
