--- web/src/components/MenuCard2/MenuCard2.test.js
@@ -1,11 +1,0 @@
-import { render } from '@redwoodjs/testing'
-
-import MenuCard2 from './MenuCard2'
-
-describe('MenuCard2', () => {
-  it('renders successfully', () => {
-    expect(() => {
-      render(<MenuCard2 />)
-    }).not.toThrow()
-  })
-})
