--- web/src/components/OrderCard/OrderCard.stories.js
@@ -1,7 +1,0 @@
-import OrderCard from './OrderCard'
-
-export const generated = () => {
-  return <OrderCard />
-}
-
-export default { title: 'Components/OrderCard' }
