--- web/src/components/OrderCard/OrderCard.test.js
@@ -1,11 +1,0 @@
-import { render } from '@redwoodjs/testing'
-
-import OrderCard from './OrderCard'
-
-describe('OrderCard', () => {
-  it('renders successfully', () => {
-    expect(() => {
-      render(<OrderCard />)
-    }).not.toThrow()
-  })
-})
