+++ web/src/components/UmamiMichelin/ItemCard/bubbleTea.jpeg
